# schemas/openapi.yaml
openapi: 3.1.0
info:
  title: Access Layer Google SSO API
  version: 1.0.0
  description: Shared authentication, authorization and audit API for internal tools.
servers:
  - url: https://access-layer.draftapps.it
    description: Production
  - url: http://localhost:8080
    description: Local development
security: []
paths:
...
  /v1/auth/start:
    get:
      summary: Start Google login for a registered tool
      parameters:
        - in: query
          name: tool_slug
          required: true
          schema: { type: string }
        - in: query
          name: return_url
          required: true
          schema: { type: string, format: uri }
        - in: query
          name: state
          required: true
          schema: { type: string }
        - in: query
          name: login_hint
          required: false
          schema: { type: string }
      responses:
        '302':
          description: Redirect to Google or tool callback on existing session
        '400':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
...
  /v1/auth/exchange:
    post:
      summary: Exchange one-time code for Access Layer identity and token
      security:
        - toolClientBasic: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExchangeRequest'
      responses:
        '200':
          description: Exchange successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExchangeResponse'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
...
  /v1/auth/logout:
    post:
      summary: Revoke a session or token
      security:
        - toolClientBasic: []
        - bearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                session_id: { type: string }
                refresh_token: { type: string }
      responses:
        '204':
          description: Logged out
...
  /v1/admin/tools:
    get:
      summary: List tools
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: search
          schema: { type: string }
          description: Case-insensitive search across tool slug and display name.
        - in: query
          name: status
          schema: { type: string, enum: [active, disabled, maintenance] }
        - in: query
          name: owner_email
          schema: { type: string, format: email }
      responses:
        '200':
          description: Tools list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ToolListResponse'
    post:
      summary: Create tool
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ToolCreateRequest'
      responses:
        '201':
          description: Created with one-time tool client secret
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ToolCreateResponse'
  /v1/admin/tools/{tool_id}:
    patch:
      summary: Update tool configuration
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: tool_id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ToolUpdateRequest'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ToolUpdateResponse'
    delete:
      summary: Delete a non-reserved tool registration
      description: Deletes the tool and cascades linked operational data. Audit logs are preserved with tool_id set to null and tool_slug retained.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: tool_id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Deleted tool metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ToolUpdateResponse'
        '403':
          description: Reserved access-admin tool cannot be deleted or actor lacks permission
...
components:
  securitySchemes:
    toolClientBasic:
      type: http
      scheme: basic
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    backupApiToken:
      type: http
      scheme: bearer
      description: Optional BACKUP_API_TOKEN for machine-to-machine backup export
  responses:
    Error:
      description: Standard error response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    ExchangeRequest:
      type: object
      required: [code, redirect_uri]
      properties:
        code: { type: string }
        redirect_uri: { type: string, format: uri }
    ExchangeResponse:
      type: object
      required: [access_token, token_type, expires_in, session, user, tool, grant, correlation_id]
      properties:
        access_token: { type: string }
        token_type: { type: string, const: Bearer }
        expires_in: { type: integer }
        refresh_token: { type: string }
        session:
          type: object
          properties:
            id: { type: string }
            issued_at: { type: string, format: date-time }
            expires_at: { type: string, format: date-time }
        user:
          $ref: '#/components/schemas/UserIdentity'
        tool:
          type: object
          properties:
            slug: { type: string }
            display_name: { type: string }
        grant:
          type: object
          properties:
            role: { type: string }
            permissions:
              type: array
              items: { type: string, pattern: "^[a-z0-9-]+(?::[a-z0-9-]+)+$" }
        correlation_id: { type: string }
...
    UserIdentity:
      type: object
      properties:
        id: { type: string }
        google_sub: { type: string }
        email: { type: string, format: email }
        email_verified: { type: boolean }
        hd: { type: string }
        display_name: { type: string }
...
    Tool:
      type: object
      required: [id, slug, display_name, status, allowed_return_urls, permission_keys]
      properties:
        id: { type: string }
        slug: { type: string }
        display_name: { type: string }
        description: { type: string, nullable: true }
        status: { type: string, enum: [active, disabled, maintenance] }
        owner_email: { type: string, format: email, nullable: true }
        allowed_return_urls:
          type: array
          items: { type: string, format: uri }
        permission_keys:
          type: array
          items: { type: string, pattern: "^[a-z0-9-]+(?::[a-z0-9-]+)+$" }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    ToolListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Tool'
    ToolCreateResponse:
      type: object
      required: [tool, tool_client_id, tool_client_secret]
      properties:
        tool:
          $ref: '#/components/schemas/Tool'
        tool_client_id: { type: string }
        tool_client_secret:
          type: string
          description: One-time secret value shown only in this response.
    ToolUpdateResponse:
      type: object
      required: [tool]
      properties:
        tool:
          $ref: '#/components/schemas/Tool'
    ToolCreateRequest:
      type: object
      required: [slug, display_name, allowed_return_urls]
      properties:
        slug: { type: string }
        display_name: { type: string }
        description: { type: string }
        owner_email: { type: string, format: email }
        allowed_return_urls:
          type: array
          items: { type: string, format: uri }
        permission_keys:
          type: array
          items: { type: string, pattern: "^[a-z0-9-]+(?::[a-z0-9-]+)+$" }
    ToolUpdateRequest:
      type: object
      properties:
        display_name: { type: string }
        description: { type: string, nullable: true }
        status: { type: string, enum: [active, disabled, maintenance] }
        owner_email: { type: string, format: email, nullable: true }
        allowed_return_urls:
          type: array
          items: { type: string, format: uri }
        permission_keys:
          type: array
          items: { type: string, pattern: "^[a-z0-9-]+(?::[a-z0-9-]+)+$" }
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, correlation_id]
          properties:
            code: { type: string }
            message: { type: string }
            correlation_id: { type: string }
            details: { type: object }